pyproject.toml
[project]
name = "django-project"
version = "0.1.0"
description = "Django project"
requires-python = ">=3.10"
dependencies = [
    "django",
    "python-decouple",
    "dj-database-url",
    "pillow",
    "whitenoise",
]

[dependency-groups]
dev = [
    "django-extensions",
    "django-debug-toolbar",
    "ruff",
]
prod = [
    "gunicorn",
    "psycopg[binary]",
]

[tool.ruff]
line-length = 120
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "DJ"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]
